Move interfaces to a common place
