+++ src/main/java/io/appium/java_client/ios/IOSDriver.java
@@ -31,6 +31,7 @@ import io.appium.java_client.InteractsWithApps;
 import io.appium.java_client.PullsFiles;
 import io.appium.java_client.LocksDevice;
 import io.appium.java_client.PerformsTouchActions;
+import io.appium.java_client.PushesFiles;
 import io.appium.java_client.SupportsLegacyAppManagement;
 import io.appium.java_client.battery.HasBattery;
 import io.appium.java_client.remote.SupportsContextSwitching;
